+++ src/utils.py
@@ -1,13 +1,10 @@
 # Utility functions
-from subprocess import PIPE, run as cmd
-from fnmatch import filter as fil
-from typing import Any
-from os import path
 
 def getRepoTags(repo_url: str):
     """
     Get the tags of a remote repository.
     """
+    from subprocess import PIPE, run as cmd
     res = cmd(["git", "ls-remote", "--tags", repo_url], stdout=PIPE, text=True)
 
     out_lines = res.stdout.splitlines()
@@ -23,6 +20,7 @@ def getRepoTagHashes(repo_url: str):
     """
     Get the hashes for the tagged releases of a remote repository.
     """
+    from subprocess import PIPE, run as cmd
     res = cmd(["git", "ls-remote", "--tags", repo_url], stdout=PIPE, text=True)
 
     out_lines = res.stdout.splitlines()
@@ -43,6 +41,7 @@ def getRepoHeads(repo_url: str):
     """
     Get the HEAD commits of a remote repository.
     """
+    from subprocess import PIPE, run as cmd
     res = cmd(['git', 'ls-remote', '--heads', repo_url], stdout=PIPE, text=True)
 
     out_lines = res.stdout.splitlines()
@@ -61,6 +60,7 @@ def getRepoHeadHash(repo_url: str):
     """
     Get the hash for the HEAD commit of a remote repository.
     """
+    from subprocess import PIPE, run as cmd
     res = cmd(['git', 'ls-remote', repo_url, 'HEAD'], shell=False, stdout=PIPE,
             text=True)
      
@@ -77,6 +77,9 @@ def include_patterns(*patterns):
     """
     Define the patterns to include in the tree when using shutil.copytree().
     """
+    from fnmatch import filter as fil
+    from os import path
+    from typing import Any
     def _ignore_patterns(p: Any, names: list[str]):
         keep = set(name for pattern in patterns
                    for name in fil(names, pattern))
@@ -88,3 +91,18 @@ def include_patterns(*patterns):
     return _ignore_patterns
 
 
+def remove_empty_dirs(dir):
+    """
+    Remove empty directories from a given base directory.
+    """
+    from os import walk, listdir
+    from shutil import rmtree
+    dirs = [x[0] for x in walk(dir)][1:]
+    dirs.reverse()
+    for sub_dir in dirs:
+        if len(listdir(sub_dir)) == 0:
+            try:
+                rmtree(sub_dir)
+                print(f"Removed {sub_dir}")
+            except Exception as err:
+                print(f"Could not delete {sub_dir} due to {err}.")